+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix redirect on missing root page
 - Don't show inactive brands
 - Fix order shipping address
+- Fix ship-to address on order payment
+- Blank duplicate postbox
+- Disable shipping on identical to invoice
 
 ### Added
 - CHANGELOG.md